+++ src/joiner/double/int/DoubleIntJoiner.cpp
@@ -12,20 +12,18 @@ namespace Mefodij {
         validate(op);
 
         if (op == L"+") {
-            left->setDouble((double) left->getDouble() + right->getInt());
+            left->setDouble(left->getDouble() + (double) right->getInt());
         } else if (op == L"-") {
-            left->setDouble((double) left->getDouble() - right->getInt());
+            left->setDouble(left->getDouble() - (double) right->getInt());
         } else if (op == L"*") {
-            left->setDouble((double) left->getDouble() * right->getInt());
+            left->setDouble(left->getDouble() * (double) right->getInt());
         } else if (op == L"/") {
             if (right->getInt() == 0) {
                     throw runtime_error("Division by zero.");
                 }
-            left->setDouble((double) left->getDouble() / right->getInt());
+            left->setDouble(left->getDouble() / (double) right->getInt());
         }  else if (op == L"%") {
-            left->setDouble(
-                (double) ((int)left->getDouble() % right->getInt())
-            );
+            left->setInt((int) left->getDouble() % right->getInt());
         } else if (op == L"==") {
             left->setBool(
                 left->getDouble() == right->getInt()
